LayoutTests/loader/navigation-policy/should-open-external-urls/subframe-click-target-top.html

<script>

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpPolicyDelegateCallbacks();
    testRunner.waitUntilDone();
}

function runTest()
{
    window.frames[0].location = "resources/iframe-click-notify-done-target-top.html";
}

</script>
<body onload="runTest();">
This page has a subframe with a link.<br>
Then a link is clicked in that subframe that navigates the main frame, and that navigation should have the "should open external urls" flag set to true because it was from a user gesture.<br>
<iframe></iframe>
</body>
